Time: 29 ms (86.69%) | Memory: 56.5 MB (84.88%) - LeetSync

+++ 1966-frequency-of-the-most-frequent-element/frequency-of-the-most-frequent-element.java
@@ -3,7 +3,7 @@ class Solution {
         long sum=0;
         int l=0,r,maxfreq=0;
         Arrays.sort(nums);
-        long n=nums.length;
+        int n=nums.length;
         for(r=0;r<n;r++)
         {
             sum+=nums[r];
